+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog
 ------------------
 - Added a new endpoint to get booking details. (#40442).
   [daniele]
+
+- Add autoconfirm content rule to profile.
+  [foxtrot-dfm1]
 - Added field "cosa_serve" (#40445).
   [daniele]